+++ src/main/scala/com/github/havarunner/exception/MultipleInvalidations.scala
@@ -1,0 +1,5 @@
+package com.github.havarunner.exception
+
+class MultipleInvalidations(val exceptions: Seq[Exception]) extends RuntimeException(
+  exceptions.mkString("\n")
+)