+++ changelog.md
@@ -2,6 +2,11 @@
 
 This project uses [Semantic Versioning](http://semver.org)
 
+## HEAD
+
+* Report multiple invalidations at once
+* Reduce noise in suite reports
+
 ## 0.10.6
 
 * Improve speed by rewriting the scheduler as non-blocking